Improve span rotation logic
This commit add a logic to avoid rotating span if there is a milestone consensus that reaches a minimum threshold but not the 67% threshold.
